core: replace the id field of an Contract with an iv field
The name `id` caused a lot of confusion. People expected it to be the identifier of an Contract which is it not.
The Contract hash is the identifier of a Contract not the `id`

This confusion forced a re-evaluation of the `id` field.

The ID field of a Contract was used for three reasons

1. Allow the creation Contracts with the same content in case of an accidental revocation/rejection
2. Sorting Contracts as the ID was a UUID v7
3. Creating Grant hashes that unique for a specific Contract

The result of the re-evaluation is to replace the `id` field with an initialization vector field named `iv`.
An initialization vector is a unique randomly created value, selected in a nonpredictaple way, which can only be used for a single Contract.
The function of the 'iv' field is to ensure Grant hashes that only reference a single Contract.

As for the other 2 reasons the `id` field was added; The sorting can be done using the `created_at` and `content_hash` of a Contract.
In case of an accidental revocation/rejection the `created_at` which contains the creation date of the Contract will cause the content hash to be different even if the content is the same.
The only caveat is that the new Contract has to be created at least a second later because the `created_at` is an Unix timestamps which is acceptable.