Add extra decision to the list when defining a Group
Since a Group can only be used in a single network, it makes sense to add this information to the rules & restrictions of a Group.
This makes it easier for users to understand where a Group is located.
Eg. public internet or DigiNetwerk.

When this MR is merged, the Digikoppeling Koppelvlakstandaard REST-API needs to be updated with the network being used.
See https://gitdocumentatie.logius.nl/publicatie/dk/restapi/#federated-service-connectivity-standaard-fsc