introduce distribute-only-peer
A Distribute-Only-Peer is a Peer who will only send requests but is unable to receive them.
This Peer is introduced for the usecase of organizations who will want to participate as Delegator
but are missing the IT knowledge to deploy an entire FSC setup.
By adding a Peer that will only send requests new implementation possibilities are created.
For example a mobile app for signing contracts.
This is a breaking change because Peers will have to send the entire certificate when
signing Contracts instead of only the certificate thumbprint.
This is needed because Distribute-Only-Peers are unable to provide key material through a JWKS endpoint.