improve content hash description
The IV field of a Contract contains an UUIDv7 which is serialized as a string.
This string must first be parsed as an UUIDv7 and the bytes of the UUIDv7 should be added to the content/grant hash.
This not realy clear when reading the description which can cause the mistake of converting the UUIDv7 string to bytes.
Using the bytes of the string will create an incorrect hash