allow outways based on domain name
In a (Delegated) Service Connection Grant an Outway can be specified
based on domain name. Previously only a public key thumbprint was
allowed. An Outway must have the domain name specified as a Subject
Alternative Name in its certificate to be allowed to request an access
token. By allowing domain names Contracts won't have to be renewed when
the Outway renews its certificate and rotates its keypair.